bodyprog: add `view` matches

+++ include/bodyprog/math.h
@@ -4,7 +4,7 @@
 void func_80096C94(SVECTOR*, MATRIX*); // custom vwRotMatrix*** ?
 void func_80096E78(SVECTOR*, MATRIX*); // another custom vwRotMatrix*** ?
 
-s32 shRcos(s16);
-s32 shRsin(s16);
+s32 shRcos(s32);
+s32 shRsin(s32);
 
 #endif /* _BODYPROG_MATH_H */
